Share selector prevalidation contract
